Introduce FeatureBarcodeMatrix.
