+++ Makefile
@@ -2,7 +2,7 @@ CXX=g++
 CXXFLAGS=-std=c++11 -Wall -O3 -fopenmp -msse4.1
 LDFLAGS=-lm -lz
 
-cpp_source=sequence_batch.cc index.cc candidate_processor.cc ksw.cc output_tools.cc chromap.cc
+cpp_source=sequence_batch.cc index.cc candidate_processor.cc feature_barcode_matrix.cc ksw.cc output_tools.cc chromap.cc
 src_dir=src
 objs_dir=objs
 objs+=$(patsubst %.cc,$(objs_dir)/%.o,$(cpp_source))
